+++ CHANGELOG.md
@@ -1,13 +1,21 @@
 ### HEAD
 
+### HEAD
+
+### 0.1.2 (October 23, 2015)
+
+  * Updated the email address,
+  * Updated the test file.
+
+
 ### 0.1.1 (October 16, 2015)
 
   * Updated the README.md file,
   * Upgraded the test file to download GeoLite2 databases if they do not exist (Travis CI),
   * Completed .eslintrc stylistic rules,
   * Fixed stylistic issues,
-  * Reworked test file,
-  * Improved test coverage.
+  * Reworked the test file,
+  * Improved the test coverage.
 
 
 ### 0.1.0 (October 9, 2015)